Add plookup tables
Insert hash tables for
T1, T2, T3
