+++ src/plookup/table/hash_tables/constants.rs
@@ -10,7 +10,7 @@
 // These are the required constants for the
 // Currently making the s_i usize.
 
-use crate::plookup::table::hash_tables::{HashTableTwo, HashTableThree};
+use crate::plookup::table::hash_tables::{HashTableThree, HashTableTwo};
 use crate::prelude::BlsScalar;
 
 /// This is the smallest prime that exceeds the BLS construction
@@ -28,438 +28,3 @@ pub const S: [u64; 27] = [
 
 /// Arity of hash table
 pub const T_S: u32 = 4;
-
-/// Hash table containing fixed binary
-/// possibilities for Hash
-pub const HASH_TABLE_TWO: HashTableTwo = HashTableTwo([
-    [
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar::zero(),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar::zero(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::zero(),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar::zero(),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::zero(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::zero(),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-    ],
-]);
-
-/// Hash table containing fixed binary
-/// possibilities for Hash
-pub const HASH_TABLE_THREE: HashTableThree = HashTableThree([
-    [
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::zero(),
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-    ],
-    [
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-    ],
-    [
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-    ],
-    [
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-    ],
-    [
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar::one(),
-    ],
-    [
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-        BlsScalar([
-            17179869180,
-            12756850513266774020,
-            3681868479150465002,
-            3479420709561305823,
-        ]),
-    ],
-]);
