+++ src/plookup/table/hash_tables/constants/third_table.rs
@@ -1,0 +1,343 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+//
+// Copyright (c) DUSK NETWORK. All rights reserved.
+
+/// This file contains a hard-coded version of the 
+/// third constraint table required in the zelbet 
+/// hash function.
+
+/// Hash table containing fixed binary
+/// possibilities for Hash
+pub const HASH_TABLE_THREE: HashTableThree = HashTableThree([
+    [
+        BlsScalar::zero(),
+        BlsScalar::zero(),
+        BlsScalar::zero(),
+        BlsScalar::zero(),
+    ],
+    [
+        BlsScalar::zero(),
+        BlsScalar::zero(),
+        BlsScalar::zero(),
+        BlsScalar::one(),
+    ],
+    [
+        BlsScalar::zero(),
+        BlsScalar::zero(),
+        BlsScalar::one(),
+        BlsScalar::one(),
+    ],
+    [
+        BlsScalar::zero(),
+        BlsScalar::zero(),
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+    ],
+    [
+        BlsScalar::zero(),
+        BlsScalar::one(),
+        BlsScalar::one(),
+        BlsScalar::one(),
+    ],
+    [
+        BlsScalar::zero(),
+        BlsScalar::one(),
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+    ],
+    [
+        BlsScalar::zero(),
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+    ],
+    [
+        BlsScalar::zero(),
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+    ],
+    [
+        BlsScalar::one(),
+        BlsScalar::one(),
+        BlsScalar::one(),
+        BlsScalar::one(),
+    ],
+    [
+        BlsScalar::one(),
+        BlsScalar::one(),
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+    ],
+    [
+        BlsScalar::one(),
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+    ],
+    [
+        BlsScalar::one(),
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+    ],
+    [
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+        BlsScalar::one(),
+    ],
+    [
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+    ],
+    [
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+    ],
+    [
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+    ],
+    [
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+        BlsScalar::one(),
+        BlsScalar::one(),
+    ],
+    [
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+    ],
+    [
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+    ],
+    [
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+    ],
+    [
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+        BlsScalar::one(),
+    ],
+    [
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+    ],
+    [
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar::one(),
+    ],
+    [
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+        BlsScalar([
+            17179869180,
+            12756850513266774020,
+            3681868479150465002,
+            3479420709561305823,
+        ]),
+    ],
+]);
